checking

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -4,7 +4,7 @@
 * @a: variable
 * Return: counter
 */
-int counting(int a)
+int counting(unsigned long int a)
 {
 	long unsigned int counter = 0;
